+++ backend/bankingSystem/src/main/java/com/example/demo/controller/AccountController.java
@@ -12,7 +12,9 @@ import com.example.demo.exception.UserNotFoundException;
 import com.example.demo.model.Account;
 import com.example.demo.model.ApiResponse;
 import com.example.demo.model.Login;
+import com.example.demo.model.User;
 import com.example.demo.repository.AccountRepository;
+import com.example.demo.repository.UserRepository;
 
 @RestController
 public class AccountController {
@@ -20,6 +22,8 @@ public class AccountController {
 	@Autowired
 	private AccountRepository accountRepository;
 	
+	@Autowired
+	private UserRepository userRepository;
 	@PostMapping("/netbanking/login")
 	public String login(@RequestBody Login login) {
 		Optional<Account> account = accountRepository.findByUsernameAndLoginPasswd(login.getUsername(), login.getPassword());
@@ -41,5 +45,26 @@ public class AccountController {
 		return "Registration for Internet Banking is successful";
 	}
 	
+	@PostMapping("/netbanking/forgot_username")
+	public String forgotUsername(@RequestBody Account account) {
+		Optional<Account> currentAccount = accountRepository.findById(account.getAccountNum());
+		if(currentAccount.isPresent()) {
+			return "123456";
+		}
+		return "Account does not exists";
+	}
 	
+	@PostMapping("/netbanking/otp_username")
+	public String confirmUsername(@RequestBody Account account) {
+		return "username";
+	}
+	
+	@PostMapping("/netbanking/forgot_password")
+	public String fogotPassword(@RequestBody Account account) {
+		Optional<User> user = userRepository.findById(account.getUser().getId());
+		if(user.isPresent()) {
+			return "123456";
+		}
+		return "user does not exists";
+	}
 }
